Rationalize treatment of non-array scalar values in PyTreeCheckpointer.
The root problem is that Tensorstore cannot distinguish between 0d arrays and non-array scalars.

With this change, we will convert 0-d arrays deserialized using tensorstore from np.ndarray to Python scalars using `item()`. We will ensure that values serialized/deserialized with flax are not forcibly converted to a different object.

PiperOrigin-RevId: 451021087