orbax: Fix host_local_to_global_array checks
This also renames the function to identical_host_local_to_global_array
to emphasize that the the shards of the array must be identical.

The input array to this utility method is normally not fully replicated,
as the the output from a pmap-ed step is usually completely sharded.

Instead of checking `is_fully_replicated` we insted make sure that the
array uses all local devices and is fully addressible. This won't catch
all misuses, but should catch most.

PiperOrigin-RevId: 530767103