Orbax: Refactor tensorstore spec construction and support local temporary index path
This separates the KvStore spec construction from the array spec construction, and integrates the options only used for writing with the main spec construction.

The support for a local temporary index path will be used in a later commit to avoid writing per-host metadata to persistent storage only to re-read and merge it.

PiperOrigin-RevId: 641574329