Add a new helper function to construct TensorStore specs for writing arrays.
`get_json_tspec_write` will be replaced with it in separate subsequent changes, and then removed. This is done to keep the code change with the actual switch small.

The only functional change in the current CL, affecting `get_json_tspec_write`, is setting `target_data_file_size` to 0 in KvStore spec if it is explicitly requested.

PiperOrigin-RevId: 676057672