Add additional support for TENSORSTORE_GCS_BACKEND environment variable
This allows the user to configure the Tensorstore GCS backend (i.e. with `gcs` for http or `gcs_grpc` for grpc) for additional places in the code that did not previously have support, including when using ocdbt.

This also refactors some previously duplicated code to parse the gcs bucket/path and generate the kvstore spec into the existing gcs_utils.py file.

PiperOrigin-RevId: 809285436